+++ Pandora/Paradigm/Inventory.hs
@@ -8,17 +8,16 @@ import Pandora.Paradigm.Inventory.Some as Exports
 import Pandora.Core.Functor (type (>>>))
 import Pandora.Core.Interpreted (run, (<~))
 import Pandora.Pattern.Semigroupoid ((.))
-import Pandora.Pattern.Category ((<--), (<---), (<----))
+import Pandora.Pattern.Category ((<--), (<---))
 import Pandora.Pattern.Kernel (constant)
-import Pandora.Pattern.Morphism.Flip (Flip (Flip))
-import Pandora.Pattern.Functor.Covariant (Covariant ((<-|-), (<-|-|-), (<-|----)))
+import Pandora.Pattern.Functor.Covariant (Covariant ((<-|-)))
 import Pandora.Pattern.Functor.Semimonoidal (Semimonoidal (mult))
 import Pandora.Pattern.Functor.Adjoint (Adjoint ((-|), (--|), (|-), (|--), (|---)))
 import Pandora.Paradigm.Primary.Functor.Maybe (Maybe (Just, Nothing))
 import Pandora.Paradigm.Primary.Functor.Exactly (Exactly)
 import Pandora.Paradigm.Algebraic.Product ((:*:) ((:*:)))
 import Pandora.Paradigm.Algebraic.Exponential ((%), type (<--))
-import Pandora.Paradigm.Algebraic (Pointable, extract)
+import Pandora.Paradigm.Algebraic.Functor (Pointable, extract, (<<-|-))
 import Pandora.Paradigm.Controlflow.Effect.Transformer ((:>) (TM))
 import Pandora.Paradigm.Controlflow.Effect.Adaptable (Adaptable (adapt))
 import Pandora.Paradigm.Schemes.TUT (TUT (TUT))
@@ -42,7 +41,7 @@ zoom :: forall bg ls u result . Lens u bg ls -> State (u ls) result -> State bg 
 zoom lens less = State <-- \source -> restruct |--- run <-- lens <~ source where
 
 	restruct :: (u ls -> bg) -> u ls -> bg :*: result
-	restruct to target = run @(->) <---- to <-|- Flip <-- less <~ target
+	restruct to target = to <<-|- less <~ target
 
 -- TODO: Use <<-|-|- instead of run <-|- ... Flip <-|- ..
 magnify :: forall bg ls t u result . Covariant (->) (->) t
@@ -50,7 +49,7 @@ magnify :: forall bg ls t u result . Covariant (->) (->) t
 magnify lens less = TM . TUT <-- \source -> restruct |--- run <-- lens <~ source where
 
 	restruct :: (u ls -> bg) -> u ls -> t (bg :*: result)
-	restruct to target = run @(->) <-|---- to <-|-|- Flip <-|- less <~ target
+	restruct to target = (to <<-|-) <-|- less <~ target
 
 overlook :: (Covariant (->) (->) t, Semimonoidal (<--) (:*:) (:*:) t) => State s result -> State (t s) (t result)
 overlook (State state) = State <-- \ts -> mult @(<--) @(:*:) @(:*:) <~ (state <-|- ts)