+++ CHANGELOG.md
@@ -792,5 +792,6 @@
 * Define `until` method for repeating actons until failure
 * Define `definitely` method to simplify `State` transformer
 * Define `magnify` method to `zoom` states in `State` transformers
-* Rename `<-||-` lenght encoding operators to `<<-|-`
+* Rename `<-||-` length encoding operators to `<<-|-`
 * Define  priority and fixity for `-+` length encoding operators
+* Define `<<-|-|-` length encoding operators